requirements.txt
click>=6.0
pyserial>=3.0
bincopy>=16.0.0
easy_enum>=0.1.0
pyusb>=1.0.0b2; platform_system == 'Linux'
pywinusb>=0.4.0; platform_system == 'Windows'
